add BaseId
New lua function: vocation:getBaseId()
New c++ function vocation->getBaseId()

The idea of the base id is that the "promotions" have the same id as the base vocation, so that it is not necessary to have repetitions of ifs and if a new vocation is added, it will also not be necessary to register them in the scripts that call the vocations , because the id is the same. This function should only be used in scripts that will be used by all base vocations.

For example, if a spell can be buy by the "knight" or "elite knight", then you can use vocation:getBaseId()
It is always preferable to use vocation:getBaseId() when possible.