Simplify imbument check logic
We keep track of how many active imbuements a player have, checking imbuement duration only for items equiped on those players, decreasing it accordingly.

When an imbuement ends, it's decreased from player active imbuements count.
When a player logs in, we run a full check initializing active imbuements for this player.
When a player is removed, we remove him from the active imbuements map.
When an item is equipped/deequipped we add/remove it's imbuements to the holder active imbuements.