[Enhancement] rework for enable choosing of a custom datapack name
This modification allows us to choose the name of the folder "datapack", to work more easily with the global pack.

If everything goes as planned, we can migrate the global repository here, without necessarily affecting the development of the custom. Who wants to use the global datapack, just change the name, who wants to use the usual "data", can also.

Also removed the c++ store functions as it was not used.

How to test:

Take the global pack, rename it to "data-global", in the src/core.hpp file, in "DATAPACK_FOLDER_NAME" put the same folder you renamed, compile and test.

Fix for not remove items with decay added from map