[Enhancement] adjust of monster (and others) speed multiplication (#485)
With this modification, we will no longer need to send the speed of monsters, items and other things, doubled.

Several bugs have already been caused because sometimes when updating a new system, we don't realize that it needs to be sent folded to make the division in the protocol. We removed this once and for all, thus making development easier. There is a script to convert your monsters (divide the speed) to the new system, we understand that the rest is not necessary, mount, items, etc., for being few things.

Conversor is here: https://github.com/opentibiabr/canary/files/10147673/creature.speed.conversion.zip
It can also be found in the pull request description.