Change from switch to unordered_map
Created a dictionary with std::unordered_map that maps AttrTypes_t values to attribute reading functions. That way you can remove the switch and the multiple ifs, and use just one function call to read the attribute.

Since std::function is a function wrapper class, it allows you to store any function or member function, with any signature, in a std::unordered_map. This allows you to map AttrTypes_t values to attribute reading functions simply and quickly.

Also, by using a single function to read the attributes, the code is easier to maintain since all the reading operations are concentrated in one place.