fix: setLootContainer without accessing freed memory
This change addresses an issue where the function setLootContainer and checkLootContainer was accessing a pointer after it had been freed by decrementing its reference counter. This was causing a use of memory after it was freed error. This change also moves the decrement reference counter line to the end of the function, making sure that the pointer is not accessed after it is freed.