Sonar fixes and others things
Added new functions to tools.h:
The first function, "convertToSafeInteger", is a template function that takes in two types: "ReturnValue" and "ConvertValue", as well as a value of "convertValue" of type "ConvertValue". It converts this value to the "ReturnValue" type and ensures that it falls within the range of 0 and the maximum value that the "ReturnValue" type can hold. It then returns the converted and safe value.

The second function, "getEnumClassNumber", is also a template function that takes in a single type: "EnumClass", and a reference to a value of "convertValue" of type "EnumClass". It uses the "magic_enum" library to convert the enumeration value to its underlying integral value and returns it.