improve: added clang format using GHA for code format (#775)
This implements the Clang formatter, a linter, to maintain a consistent code style throughout the code base.

In addition to configuring the Clang formatter, a GH Action has been added to automatically run the linter on the code base every time a push or pull request is made and make a commit to apply the format whenever necessary. This way, we can guarantee that the code is always in the same format.

This also allows us to merge faster, as we can simply apply the linter to any open pull request to resolve any conflicts or changes that may occur during the merge.