improve: handling of modal window helper (#794)
This adds the possibility to create modal window in a more beautiful way.

The order of execution of the calls is: 
• choice
• button
• default

If a choice has a callback it will be activated when you press any button, however the button function will not be activated even if it has one. If another choice does not have a callback, the button callback is called, finally the default callback will be called if there is no callback associated with the button or the choice.

Credits to authors are given in the pull request description.