fix: duplicated active events warning and log (#768)
Previously, having two active events for the same date would cause a warning on server load. This pull request fixes the issue and also adds a log for all active events on server load. However, be careful when defining two events with values other than 100, as it will double the final value.