fix: skills rates stages (#951)
Fixes an issue where the plain rates for magic/skills from config were not working, as the output remained the same no matter what rate was set. After investigation, it was discovered that the server was using staged configuration, even with rateUseStages set to false. Further testing was done with training/exercise weapons and weapons against monsters, confirming that setting rateUseStages were using staged. The code has been improved to fix this issue, and the PR includes updates to verify skills and magic.