feat: shortcut to rascacoon, access to the wreckoning and others fixes (#953)
• Shortcut to Rascacoon and access to The Wreckoning, mount reward on Tentugly, killing in the name of points no longer resets on freequest

This adds the ability to access Rascacoon and Wreckoning through a shortcut provided in freequests, and also adds a mount reward on Tentugly. Additionally, killing in the name of points on freequest no longer resets the points. To test these features, it is necessary to fix the door below Venore harbour, according to issue #945. Sail with Eustacio to access the shortcut and with Captain Jack Rat to access the Wreckoning.