feat: sound effects system (source code and lua core)
The sound system will be split into two commits, one for the core (required) and the other for the scripts (optional).

This commit contains the "core" part of source and lua, they are needed for the scripts with sound modifications to work correctly and are needed for the server and the sounds to work correctly.