feat: concoctions system (#1223)
Implementing a system for handling concoctions (also known as Tibiadrome potions). https://tibia.fandom.com/wiki/Tibiadrome#Rewards

The system introduces a new concoctions_lib.lua file that contains the majority of the logic. This file provides a registry to create concoction configurations and allows for later inspection of the state and duration of the concoctions for a player.

Additionally, a new configuration key called partyShareLootBoosts has been added. This key enables any party member to provide a loot boost to the entire party through either prey or wealth duplex. This functionality deviates from how Tibia normally operates, but it has been included as an option to address the annoyance associated with the default Tibia behaviour.

Please note that this implementation only includes four of the available concoctions. The remaining concoctions can be implemented at a later stage.