improve: move chaining to cpp (#1244)
Like the title said, moves chaining to cpp. This makes it easier/quicker to implement chaining spells (such as divine dazzle and chivalrous challenge) but also makes it easier to make mobs use chaining attacks (Like this PR does for Harpy and Carnivostrich).

This could also have been a lib in lua, but move the mechanic to c++ since it's a core part of Tibia these days and not just something that is content specific.