fix: use loot modifiers correctly when creating loot (#1310)
Container:createLootItem now returns false when no loot was created (either by error or because we rolled a random that didn't add it). This was causing a warning to pop up on the monster callback. This was a harmless warning, but we were missing logs when something actually went wrong (not enough space on the corpse or invalid itemID).

This adds logs inside createLootItem and silences the spurious alerts above.