feat: Introduces Dependency Injection (DI) via boost DI
Dependency Injection (DI) is a powerful technique used to decouple classes from its dependencies. This improves our ability to isolate and encapsulate code, as well as centralise and standardise the definition of dependencies.

This commit introduces dependency injection using the BoostDI library and an empty template of a container, to be extended and used in future commits.