feat: Abstracts logging into ILogger interface
Currently we use SPDLOG for logging. However, we use SPDLOG macros directly all over the place. Although it's quite handy, it makes our code very attached to SPDLOG and hard to abstract logging away for tests and others.

Aiming to improve that, we are abstracting logging into an interface ILogger and adding a SpdLog implementation of this interface.

We are also defining SpdLog as the default implementation of ILogger in our DI, so that everything using DI to inject logger will always get SpdLog as default.