fix: send only valid monsters to prey data (#1387)
# Description

When a creature with an invalid race ID is saved in the player and tries
to send it to the client, it debugs and sends the log. This way, now it
will only send the log, removes the prey slot referring to the player's
monster and send the empty bytes.