fix: wheel of destiny damage buff leaking into immunity (#1381)
Prior to this, monsters that are not supposed to that damage, such as the Giant Beaver and Magic Towers still take damage from the boost.

This is kind of a hacky change, but we rely on the damage being zero going into the boost to filter it out. It will work in all cases, and the worst thing it's doing is not giving the boost when you completely miss a hit.