feat: add "logger" lua function support for brace formatting
Introduces a new feature that allows the use of curly braces (`{}`) for variable substitution in log messages, aligning the formatting with the standard C++ convention.

With this change, the "Spdlog" object is deprecated, and we have added the new "logger" object. The main difference between the two is that the new "logger" supports advanced formatting using braces, allowing greater flexibility and control over log messages, just as it is done in C++.

The new logging system, which will replace "Spdlog." It supports advanced formatting using braces.
New functions:
logger.info("")
logger.warn("")
logger.error("")
logger.debug("")

Example of use:
local string = "formatting"
logger.info("Information message with {}", string)

Output:
[2023-21-08 05:43:46.765] [info] Information message with formatting

It is highly recommended to migrate to the new "logger" system to take advantage of the advanced formatting features and ensure compatibility with future updates. The use of "Spdlog" is still possible, but it is deprecated, and its use is discouraged.

The change includes:
• Support for replacing strings, numbers, userdata, and other types in Lua within log messages.
• The extraction of specific functionalities into separate functions, such as converting the userdata type into its corresponding name.
• The correction of any issues identified in forming the log string.
• The use of the "magic enum" library to simplify enum conversion, providing an elegant and efficient way to translate between enums and strings.
• Added logger example in "bestiary_kill" and "bosstiary_kill" scripts, and added sanity checks.

This enhancement will make it easier to write more readable and consistent log messages, adhering to common practices in C++ and providing a better development experience.