improve: trigger source builds only when source files are modified (#1475)
Enhances the efficiency of our build process by triggering builds only when changes are made to the source files. This selective triggering applies to various workflows including:

• Windows Build: Both CMake and Solution configurations are affected.
• Ubuntu Build: The build process on the Ubuntu platform has been updated to follow this new rule.
• SonarCloud: The integration with SonarCloud will now only run when relevant source files are modified.
• Docker: The Docker build workflow has been optimized to only execute when necessary.

By focusing on changes to the source files, this improvement ensures that unnecessary builds are avoided, saving both time and resources. It aligns with our commitment to an efficient and streamlined development process.

NOTE: I also added an option that will cancel old (concurrent) branch builds if a new commit is submitted.