feat: Use runtime_provider DI and add tests
runtime_provider allow us to add dependencies on the fly. While not a good practice in production - once created we want our DI to be as static and reliable as possible - it's very useful for tests.

Installing dependencies on the fly for tests and overriding production setup allow us to test real flows with stubs, removing the need of testing with mocks and improving our test range/power.