improve: Tweak Lua shutdown to avoid SEGFAULT
LuaInterface is really badly coupled to ScriptInterface and other parts of the code that integrate with lua. It has some cross-references and cross-dependencies, and this makes shutdown quite tricky.

Because they depend on each other, we need to avoid calls to one another and reduce the scope of the destructors to make sure that the application can safely die.