fix: inject g_logger from interface not impl
Logger interface was created to generate testable abstraction to tests. This is working fine for log ingestion in construction level. However, for those using g_logger it has a hard bind to the impl class LogWithSpdLog instead of the Interface.

This commit changes this to make it more flexible and improve testing.