improve: change shared_ptr from reference to value (#1506)
It appears that passing some shared_ptrs by reference was causing
crashes, as the reference was deleted before being used. I've reverted
most of these instances that seemed likely to cause problems, but I've
left a few in place for reference, such as in the case of Raid and
GlobalEvent, which seem less prone to issues.