fix: webhook race condition crash
When sending webhooks there is a small chance that the webhooks.empty() if passes but when it gets the task the webhooks deque is already empty.

That causes task to be a nullptr and accessing it causes a segmentation fault.

There are two ways to solve if: place the lock in the beginning of the function or check if task exists before proceeding, we are adopting the first.