feat: more robust zone handling (#1524)
This adds a lot of safety to how we handle Zones. Making it so dealing
with the contents of a zone is never going to respond with creatures
that have already been deleted.

This also migrated the `bosslever.lua` to use zones, and deprectes the
Spectators library. Zones are a more flexible way to achieve the same
thing, and more. This should also fix a potential crazy with the removal
of players from a bosslever room since that was previously done with an
unsafe call to `addEvent`.