feat: chain improvements for lower range (#1554)
The C++ chain system works well today if the chain ability can hit
further than 2 squares. But for 1 tile chaining it tends to pick too few
targets and be sub-optimal. This isn't a big issue with the current
spells/monsters in the datapack but it poses a limitation to developing
chain abilities with shorter range.
This PR fixes that by re-working the chain algorithm to do more
efficient backtracking in order to pick a good set of targets to hit.