feat: item ownership (#2003)
This add the ability to set ownership on items. An owned item cannot be
picked up or used by anyone except the character that owns it.

If an owned item is inside a container (backpack, etc) that container is
now also treated as owned. "looking" at an owned item shows you the
owner.

New function:
- `item:setOwner(ownerPlayer or id)`
- `item:getOwnerId()`
- `item:getOwnerName()`
- `item:hasOwner()`
- `item:canBeMoved()`

I also sneaked in some other unrelated improvements that didn't fit any
individual PR in here, but the are pretty small.

---------

Co-authored-by: Eduardo Dantas <eduardo.dantas@hotmail.com.br>