fix: cobras spawn and removal invalid scripts (#2264)
Remove the event from the main callback and move it inside the action.
Also, remove the globalevent that recreates the cobra flask because
there is no information that this is the default behavior, and according
to a friend who plays the global, he said it can only be picked up once
per ss.