Fix uptime
There is an uptime error in protocolstatus, as it does not load the time in the cache correctly, causing it to not update the time.

I added this variable within uptime, so that, instead of using the cache time, in specific cases it loads the system time, so it is possible to use this in other parts where uptime is a problem, just adding the true inside OTSYS_TIME().