fix: supreme revelation modifiers initialization
This fixes several issues causing the "Supreme Revelation Perks" from the wheel to not work correctly.

Firstly, the `m_value` in `wheel_gems.hpp` was not being initialized, resulting in the extra points value always being 0.
Secondly, the order of variable initialization was incorrect, which caused the points to be used before they were loaded, also resulting in a value of 0.

Additionally, smaller functions were created to improve readability and maintainability, thus preventing future errors like this one.