fix: resolve nil value errors in handleGuildWar function (#3172)
I made corrections to the `death.lua` script to resolve nil value errors
in the `handleGuildWar` function. I added checks to ensure that both the
player and the killer are valid and that both belong to a guild before
proceeding with the guild war logic. This helps prevent failures when
one of the objects is undefined.

Fixes from: https://github.com/opentibiabr/canary/commit/3c98b4161cd9c25e57e61945fe6358d604242b27