fix: necklace/ring slots exhaustion (#3272)
This extends #3165 in the following manner:
I've separated necklace and ring slots in different groups, so the `ACTIONS_DELAY_INTERVAL` delay flag works on the groups separately. 
This means the exhaustion delay continues to be applied on equip items, but considering groups separately. In other words, the players are now able to equip necklaces and rings at the same time, but when trying to equip different necklaces (or rings) they are subject to `ACTIONS_DELAY_INTERVAL`.