/setskill level, correction.
corrected the well pointed level leak where the level and XP would be changed correctly but health, mana and capacity stats stay the same.
tested every scenario and all seems to be working fine, please review.

Also I don't think the code should be separated in two files because it's the same thing with two different approaches, there's no need to split and duplicate code/files... and performance wouldn't be an issue since the script is only triggered on god talkaction, it's not a big deal.