fix: sendUpdateTileCreature log client debug (#3748)
Resolves: #3741 

This change makes a targeted improvement to the
"sendUpdateTileCreature" function in "protocolgame.cpp". The main change
ensures that the "known" status of a creature is correctly passed to the
"AddCreature" function, which likely affects how creature updates are
handled in the game protocol.

- Networking and game protocol:
* In `ProtocolGame::sendUpdateTileCreature`, initialized `removedKnown`
to 0 and updated the call to `AddCreature` to pass the `known` variable,
ensuring accurate handling of whether a creature is already known to the
client.